+++ client/lib/pages/individual_challenge_page.dart
@@ -1,138 +1,246 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
+import 'package:geofence_service/geofence_service.dart';
 import 'package:trailquest/widgets/back_button.dart';
 import 'package:trailquest/widgets/challenge.dart';
 
 class IndividualChallengePage extends StatefulWidget {
   Challenge challenge;
 
-  IndividualChallengePage({
-    required this.challenge
-  });
+  IndividualChallengePage({required this.challenge});
 
   @override
-  State<IndividualChallengePage> createState() => _IndividualChallengeState(challenge: challenge);
+  State<IndividualChallengePage> createState() =>
+      _IndividualChallengeState(challenge: challenge);
 }
 
 class _IndividualChallengeState extends State<IndividualChallengePage> {
+  // Controllers for geofences
+  final _activityStreamController = StreamController<Activity>();
+  final _geofenceStreamController = StreamController<Geofence>();
+
+  // Settings for geofences
+  final _geofenceService = GeofenceService.instance.setup(
+      interval: 2000,
+      accuracy: 10,
+      loiteringDelayMs: 60000,
+      statusChangeDelayMs: 10000,
+      useActivityRecognition: true,
+      allowMockLocations: true,
+      printDevLog: false,
+      geofenceRadiusSortType: GeofenceRadiusSortType.DESC);
+
+  // Automatically called every time the status of a geofence is changed
+  // Most game logic will be placed here
+  Future<void> _onGeofenceStatusChanged(
+      Geofence geofence,
+      GeofenceRadius geofenceRadius,
+      GeofenceStatus geofenceStatus,
+      Location location) async {
+    print('geofence: ${geofence.toJson()}');
+    print('geofenceRadius: ${geofenceRadius.toJson()}');
+    print('geofenceStatus: ${geofenceStatus.toString()}');
+    _geofenceStreamController.sink.add(geofence);
+
+    // switch (gameMode) {
+    //   // No game. Only route-generation
+    //   case 0:
+    //     break;
+    //   // Orienteering
+    //   case 1:
+    //     break;
+    //   // Checkpoints
+    //   case 2:
+    //     break;
+    //   // Treasure hunt
+    //   case 3:
+    //     break;
+    //   default:
+    //     throw Exception('Invalid gamemode');
+    // }
+  }
+
+  // Unused
+  // This function is to be called when the activity has changed.
+  void _onActivityChanged(Activity prevActivity, Activity currActivity) {
+    print('prevActivity: ${prevActivity.toJson()}');
+    print('currActivity: ${currActivity.toJson()}');
+    _activityStreamController.sink.add(currActivity);
+  }
+
+  // Unused
+  // This function is to be called when the location has changed.
+  void _onLocationChanged(Location location) {
+    print('location: ${location.toJson()}');
+  }
+
+  // Unused
+  // This function is to be called when a location services status change occurs
+  // since the service was started.
+  void _onLocationServicesStatusChanged(bool status) {
+    print('isLocationServicesEnabled: $status');
+  }
+
+  // This function is used to handle errors that occur in the service.
+  void _onError(error) {
+    final errorCode = getErrorCodesFromError(error);
+    if (errorCode == null) {
+      print('Undefined error: $error');
+      return;
+    }
+
+    print('ErrorCode: $errorCode');
+  }
 
   Challenge challenge;
 
   _IndividualChallengeState({required this.challenge});
 
   @override
+  void initState() {
+    super.initState();
+
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _geofenceService
+          .addGeofenceStatusChangeListener(_onGeofenceStatusChanged);
+      _geofenceService.addLocationChangeListener(_onLocationChanged);
+      _geofenceService.addLocationServicesStatusChangeListener(
+          _onLocationServicesStatusChanged);
+      _geofenceService.addActivityChangeListener(_onActivityChanged);
+      _geofenceService.addStreamErrorListener(_onError);
+      _geofenceService.start().catchError(_onError);
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
     String type = widget.challenge.type;
 
-    return Scaffold(
-      body: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Row(
+    return WillStartForegroundTask(
+        onWillStart: () async {
+          return _geofenceService.isRunningService;
+        },
+        androidNotificationOptions: AndroidNotificationOptions(
+          channelId: 'geofence_service_notification_channel',
+          channelName: 'Geofence Service Notification',
+          channelDescription:
+              'This notification appears when the geofence service is running in the background.',
+          channelImportance: NotificationChannelImportance.LOW,
+          priority: NotificationPriority.LOW,
+          isSticky: false,
+        ),
+        iosNotificationOptions: const IOSNotificationOptions(),
+        foregroundTaskOptions: const ForegroundTaskOptions(),
+        notificationTitle: 'Geofence Service is running',
+        notificationText: 'Tap to return to the app',
+        child: Scaffold(
+          body: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
             children: [
-              GoBackButton(), 
-              Padding(
-                padding: const EdgeInsets.only(top: 15.0),
-                child: Text(widget.challenge.name, style: TextStyle(fontSize: 25)),
-              )
-            ],
-          ),
-          Expanded(
-            child: SingleChildScrollView(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
+              Row(
                 children: [
+                  GoBackButton(),
                   Padding(
-                    padding: const EdgeInsets.only(top: 8.0),
-                    child: Image(
-                      width: 400,
-                      fit: BoxFit.contain,
-                      image: AssetImage(widget.challenge.image),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Text(
-                      widget.challenge.description,
-                      style: TextStyle(fontSize: 20),
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Container(
-                      decoration: BoxDecoration(border: Border.all(color: Colors.black)),
-                      child: Padding(
-                        padding: const EdgeInsets.all(8.0),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              'Instruction: $type',
-                              style: TextStyle(fontSize: 18),
-                            ),
-                            ChallengeInstruction(widget.challenge),
-                          ],
-                        ),
-                      ),
-                    ),
-                  ),
+                    padding: const EdgeInsets.only(top: 15.0),
+                    child: Text(widget.challenge.name,
+                        style: TextStyle(fontSize: 25)),
+                  )
                 ],
               ),
-            ),
+              Expanded(
+                child: SingleChildScrollView(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
+                    children: [
+                      Padding(
+                        padding: const EdgeInsets.only(top: 8.0),
+                        child: Image(
+                          width: 400,
+                          fit: BoxFit.contain,
+                          image: AssetImage(widget.challenge.image),
+                        ),
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: Text(
+                          widget.challenge.description,
+                          style: TextStyle(fontSize: 20),
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: Container(
+                          decoration: BoxDecoration(
+                              border: Border.all(color: Colors.black)),
+                          child: Padding(
+                            padding: const EdgeInsets.all(8.0),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                Text(
+                                  'Instruction: $type',
+                                  style: TextStyle(fontSize: 18),
+                                ),
+                                ChallengeInstruction(widget.challenge),
+                              ],
+                            ),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: TextButton(
+                  onPressed: () {
+                    setState(() {
+                      if (widget.challenge.status == 0) {
+                        widget.challenge.status = 1;
+                      } else if (widget.challenge.status == 1) {
+                        widget.challenge.status = 0;
+                      }
+                    });
+                  },
+                  style: StyleStartStopChallenge(widget.challenge),
+                  child: TextStartStopChallenge(widget.challenge),
+                ),
+              ),
+            ],
           ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextButton(
-              onPressed: () {
-                setState(() {
-                  if (widget.challenge.status == 0) {
-                    widget.challenge.status = 1;
-                  } else if (widget.challenge.status == 1) {
-                    widget.challenge.status = 0;
-                  }
-                });
-              },
-              style: StyleStartStopChallenge(widget.challenge),
-              child: TextStartStopChallenge(widget.challenge),
-            ),
-          ),
-        ],
-      ),
-    );
+        ));
   }
 }
 
-Widget ChallengeInstruction(Challenge challenge){
+Widget ChallengeInstruction(Challenge challenge) {
   String type = challenge.type;
 
-  if(type == 'Checkpoints'){
-    return Text('In a checkpoint challenge you will receive a trail with a number of checkpoints you need to visit. You need to visit all of them to complete the challenge',
-      style: TextStyle(
-        fontSize: 16
-      ),
-      textAlign: TextAlign.center
+  if (type == 'Checkpoints') {
+    return Text(
+        'In a checkpoint challenge you will receive a trail with a number of checkpoints you need to visit. You need to visit all of them to complete the challenge',
+        style: TextStyle(fontSize: 16),
+        textAlign: TextAlign.center);
+  } else if (type == 'Treasure hunt') {
+    return Text(
+        "In a treasure hunt you need to visit a number of locations to complete the challenge. You will not know where these are located beforehand.\n\n When pressing 'start challenge' you will receive a trail leading to the first location. When you have reached that location, you will receive a trail leading to the next one and so on.",
+        style: TextStyle(fontSize: 16),
+        textAlign: TextAlign.center);
+  } else if (type == 'Orienteering') {
+    return Text(
+      "In orienteering you will receive a number of control points to visit. How you get to these locations is up to you, but you must visit all control points to complete the challenge",
+      style: TextStyle(fontSize: 16),
+      textAlign: TextAlign.center,
     );
-  } else if(type == 'Treasure hunt') {
-    return Text("In a treasure hunt you need to visit a number of locations to complete the challenge. You will not know where these are located beforehand.\n\n When pressing 'start challenge' you will receive a trail leading to the first location. When you have reached that location, you will receive a trail leading to the next one and so on.",
-      style: TextStyle(
-        fontSize: 16
-      ),
-      textAlign: TextAlign.center
-    );
-  } else if(type == 'Orienteering') {
-    return Text("In orienteering you will receive a number of control points to visit. How you get to these locations is up to you, but you must visit all control points to complete the challenge",
-      style: TextStyle(
-        fontSize: 16
-      ),
-      textAlign: TextAlign.center
-    ,);
   } else {
     return Text(' ');
   }
 }
 
 ButtonStyle StyleStartStopChallenge(Challenge challenge) {
-  if(challenge.status == 0) {
+  if (challenge.status == 0) {
     return TextButton.styleFrom(
       backgroundColor: Colors.green.shade600,
       padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 80),
@@ -140,7 +248,7 @@ ButtonStyle StyleStartStopChallenge(Challenge challenge) {
         borderRadius: BorderRadius.all(Radius.circular(10)),
       ),
     );
-  } else if (challenge.status == 1){
+  } else if (challenge.status == 1) {
     return TextButton.styleFrom(
       backgroundColor: Colors.green.shade900,
       padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 80),
@@ -160,11 +268,14 @@ ButtonStyle StyleStartStopChallenge(Challenge challenge) {
 }
 
 Text TextStartStopChallenge(Challenge challenge) {
-  if(challenge.status == 0) {
-    return Text('Start challenge', style: TextStyle(color: Colors.white, fontSize: 25));
-  } else if(challenge.status == 1) {
-    return Text('Stop challenge?', style: TextStyle(color: Colors.white, fontSize: 25));
+  if (challenge.status == 0) {
+    return Text('Start challenge',
+        style: TextStyle(color: Colors.white, fontSize: 25));
+  } else if (challenge.status == 1) {
+    return Text('Stop challenge?',
+        style: TextStyle(color: Colors.white, fontSize: 25));
   } else {
-    return Text('Finished challenge!', style: TextStyle(color: Colors.white, fontSize: 25));
+    return Text('Finished challenge!',
+        style: TextStyle(color: Colors.white, fontSize: 25));
   }
 }
